feat: add gets for all queries
