+++ src/app.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import UseController from './controllers/useControllers';
+import UseController from './controllers/useGetController';
 import CommonService from './services/commonServices';
 
 @Module({
