+++ src/data/inMemoryDB.ts
@@ -1,5 +1,9 @@
 const db = {
   user: [],
+  track: [],
+  artist: [],
+  album: [],
+  favs: [],
 };
 
 export default db;
