+++ src/data/types/dataTypes.ts
@@ -33,3 +33,5 @@ interface Favorites {
   albums: string[]; // favorite albums ids
   tracks: string[]; // favorite tracks ids
 }
+
+export { User, Artist, Track, Album, Favorites };
