+++ src/services/commonServices.ts
@@ -3,7 +3,7 @@ import db from 'src/data/inMemoryDB';
 
 @Injectable()
 export default class CommonService {
-  getAll(entityType: string): object {
+  getAll(entityType: string): Array<object> {
     return db[entityType];
   }
 }
